pyproject.toml
[project]
name = "banking-intent-router"
version = "0.1.0"
description = "Banking query intent router using fine-tuned sentence embeddings and NeMo Guardrails"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "datasets>=4.8.4",
    "ipywidgets>=8.1.8",
    "peft>=0.19.1",
    "sentence-transformers>=5.4.1",
    "torch==2.5.1+cu121",
    "torchvision==0.20.1+cu121",
    "torchaudio==2.5.1+cu121",
    "nemoguardrails[nvidia]>=0.21.0",
    "python-dotenv>=1.2.2",
    "langchain-anthropic>=1.4.1",
    "langchain>=1.2.15",
    "anthropic>=0.97.0",
    "a2a-sdk>=1.0.2",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "sys_platform == 'win32'" }]
torchvision = [{ index = "pytorch-cu121", marker = "sys_platform == 'win32'" }]
torchaudio = [{ index = "pytorch-cu121", marker = "sys_platform == 'win32'" }]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv]
environments = ["sys_platform == 'win32'"]
